added logout button to offcanvas.
The height of the components is no longer stretching the page. The top nav has been made fixed to the top of the page, so vh-100 is now exactly the size of the container, rather than the size of the container plus the height of the navbar. To adjust the position of components within these containers, padding had to be introduced before the components.